Exception handling - delete

+++ src/main/java/com/developermobile/sisvenda_api/services/ProductService.java
@@ -5,8 +5,12 @@ import com.developermobile.sisvenda_api.dto.ProductMinDTO;
 import com.developermobile.sisvenda_api.entities.Client;
 import com.developermobile.sisvenda_api.entities.Product;
 import com.developermobile.sisvenda_api.repository.ProductRepository;
+import com.developermobile.sisvenda_api.resources.exceptions.DatabaseException;
 import com.developermobile.sisvenda_api.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -33,13 +37,27 @@ public class ProductService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public Product update(Long id, Product product) {
-        Product entity = repository.getReferenceById(id);
-        updateData(entity, product);
-        return repository.save(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            updateData(entity, product);
+            return repository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(Product entity, Product obj) {
